+++ src/main/java/com/kakao/saramaracommunity/member/dto/api/request/MemberLoginRequest.java
@@ -1,0 +1,13 @@
+package com.kakao.saramaracommunity.member.dto.api.request;
+
+public record MemberLoginRequest (
+	String email,
+	String password
+) {
+	public static MemberLoginRequest of(
+			String email,
+			String password
+	) {
+		return new MemberLoginRequest(email, password);
+	}
+}
