+++ src/main/java/com/kakao/saramaracommunity/member/dto/api/request/MemberRegisterRequest.java
@@ -1,0 +1,15 @@
+package com.kakao.saramaracommunity.member.dto.api.request;
+
+public record MemberRegisterRequest (
+	String email,
+	String password,
+	String nickname
+) {
+	public static MemberRegisterRequest of(
+			String email,
+			String password,
+			String nickname
+	) {
+		return new MemberRegisterRequest(email, password, nickname);
+	}
+}
