+++ src/main/java/com/kakao/saramaracommunity/member/service/MemberService.java
@@ -1,8 +1,8 @@
 package com.kakao.saramaracommunity.member.service;
 
-import com.kakao.saramaracommunity.member.controller.request.MemberLoginRequest;
-import com.kakao.saramaracommunity.member.controller.request.MemberRegisterRequest;
-import com.kakao.saramaracommunity.member.controller.response.MemberInfoResponse;
+import com.kakao.saramaracommunity.member.dto.api.request.MemberLoginRequest;
+import com.kakao.saramaracommunity.member.dto.api.request.MemberRegisterRequest;
+import com.kakao.saramaracommunity.member.dto.business.response.MemberInfoResponse;
 import com.kakao.saramaracommunity.member.entity.Member;
 
 public interface MemberService {
