+++ src/test/java/com/kakao/saramaracommunity/member/service/MemberServiceImplTest.java
@@ -1,7 +1,7 @@
 package com.kakao.saramaracommunity.member.service;
 
-import com.kakao.saramaracommunity.member.controller.request.MemberRegisterRequest;
-import com.kakao.saramaracommunity.member.controller.response.MemberInfoResponse;
+import com.kakao.saramaracommunity.member.dto.api.request.MemberRegisterRequest;
+import com.kakao.saramaracommunity.member.dto.business.response.MemberInfoResponse;
 import com.kakao.saramaracommunity.member.entity.Member;
 import com.kakao.saramaracommunity.member.repository.MemberRepository;
 import org.junit.jupiter.api.Disabled;
